gate: Collect and report stats from wdogs
So, the first constraint we have is that functions will typically
be called via ipvs direct routing which has no idea about how
many connections to which URL has been made. It has a ConnCount
thing, but when a caller re-used port the counter doesn't change.
Neither ipvs can help us with timings and other higher level things.

IOW -- the only point where we can collect stats is the wdog itself.
So the previous commit just makes this -- wraps all the API calls
into a middleware that counts things.

Since we collect stats in PODs with wdogs we need to collect them
from gate. What are the options?

1. Make wdog push stats to gate or to a database. This won't scale
   at all. Also we'll have to teach wdog authorize itself to the
   database (meh) or at gate (can be done via token, but ... won't
   scale)

2. Make gate call wdog from time to time. This is nicer, well have
   some delays in stats collecting but we can bear with it. The
   problem is that PODs die w/o our controll, and since polling
   of PODs will come with big delays for many functions, we may
   lose much more data if POD just dies.

3. Make wdog save stats somewhere and gate pick them up when it can.
   The "somehwere" should be independent from POD life time. This
   is called volumes and this is how stats collecting works.

So, wdogs put stats into a local directory which is a volume nearby
the sources (TODO: protect this dir from modification by fn code).
Gate starts a goroutine that scans the dirs and sums the stats up.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>