postgres: Initial preparations
PG doesn't recommend opening remote admin access to the db.
Istead, there's a trusted local connection to the server
available for admin. For swifty we need to remotely add
users and bases, so here's the daemon (that needs to get
https in it) for two basic ops -- create user+db and drop
the pair.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>