s3: Allow to setup cached object size
Currently if object size is less than 8M we're
putting it into mongo db instead of rados storage.

Add a handle to disable it and always use a storage.
Note: if objects are stored with one -cached-obj-size
settings and the s3 gate restarted with another settings
the objects may be not retrievable untill restarted
with old value.

Thus for developers mostly.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>