s3: Encrypt hmac secret keys in database
It's unsafe to keep secrets in plain text, we need to encrypt
them. Encryption is done with AES algo and a key located in
config (more exactly -- in secrets, referenced by config var).

For development purposes the -devel flag may be passed, in this
case the keys will be read in plain text.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>