s3: Per-bucket notifications via AMQP
Admin can subscribe particular namespace:bucket to send notifications
as AMQP messages. The vhost to send to is configured, the queue to send
to is requested when subsribing.

Missing part is throttling. AMQP unequivocally states, that published
does NOT mess with notifications/acknowleges from consumer(s), so we will
have to invent this ourselves.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>