gate: Initial changes for maria quotas
Maria does not have native quotas, so we have to invent our own. Per-DB triggers
are not possible. Fortunately, we have the way to request for db size and number
of rows, so here is the daemon that calls maria for the list of tables that
over use the quota.

TODO: lock access and unlock it back. Problem here is that existing connections
are not terminated. It is possible to force-break them, but then lambdas will
not auto-reconnect.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>