gate: Remove "build" instance
The whole idea of building a function in a resarting deployment
turned out to be very bad. Complex state maching and long delays
in updating are the biggest pains.

Let's fix this.

There should be a build deployment out there in k8s cluster where
all building runtimes build their scripts. Respectively, the pod
in it gets the whole sources shared dir, exposes "run" call outside
and builds whatever is asked immediately.

The build is thus synchronous, not sure how good (or bad) it is.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>