gate: Do not keep all crons in a single runner
This drives us into the need to keep track of cron job IDs
in the database, update this list, etc. Ugh... Instead, let's
have individual cron.Cron object for every function that needs
one and completely replacing one with a new when updating the
cron event (or just demolishing it when changing event source).

The difficulty with this approach is what key to use in the
map keeping the cron objects. Using fn.Cookie is tempting, but
while updating the event we may have two crons for one fn at
the same time :) and the map in question needs a special care.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>