s3: Make it possible to overwrite an object
When a new object gets uploaded into the s3 the previous one sits there until
the new one is activated. After this the old "version" is GC-ed.

So make this work introduce a "rover" on a bucket -- an increasing 64-bit number which
is bumped every time a new object is added and its new value it saved on the object.
The current-version of the obj is then that with the greatest number among the keys.

The FindCurObj is as fast as it was, as ocookie index is still there (though it is not
unique any more) and the find is equipped with sort+limit modifier. The latter should
be fast, as MGO docs say that output limit affects the sort behavior.

Respectively the list-objects-in-bucket is fixed -- as keys are sorted in output
we can use simple "seen previous" filter.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>