s3: Handle website requests
There are two places where we do not fit Amz.
1. There is no need in explicit policy configuration. PUT ?website will
do it for you, granting GET-s on objects for everyone.
2. The URL to read the buckets is s3.swifty.cloud:8080/ACID/bucket/...
where the ACID thing is a random ID reported in X-Swifty-Website header
when doing the GET ?website request on a bucket. For an account this
value is the same for all his/hers buckets.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>