evt: Generalize handling of /call method
Nowadays the request to /call results in looking up for a function with
URL bool set to true. This is boring and not extendable. Instead, intoduce
the notion of URL interface. The guy sits in a sync.Map and provides a
.Handle method. Once /call is requested, find the URL by ID and Handle() it.

For existing per-fn URL triggers handling means just running the FN in question.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>