mongo: Implement limiting by 4 parameters
real_size -- disk size + index size. This is the most exact, but the least
convenient limit. The problem is that disk size never goes down, so there
is no way to shrink the db down once it hits the limit.

data_size -- the amt of bytes occupied by data. It can be bigger than the
disk size if compression is in use, but it can go down when user removes
docs or drops collections.

indx_size -- size of the index. Indices are performance crucial, but occupy
space. Not sure how to deal with it though.

virt_size -- sum of data and index. Looks like this is the most sane limit
to use, as freed space from removed objects can be re-occupied by new ones
without growing the disk size.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>