swift: Fix paths on which builder puts runner
For some reason the swift build was broken for quite a long time. Builder
could not move the built binary, as it was put in another places. Runner,
at the same time, could not start the binary at the desired location, as
the debug -> symlink remained real dir non-executable and execve just failed.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>