cs: Add class for calling read/writes on file descs
This is derived from Mono.Posix code, but without preliminary empty buffer write
into the given descriptor. If we do this write, the peer end  (which is a unix
seqpacket socket) wakes up and reports EOF.

This code is not embed into runner.cs as it gets compiled with -unsafe option
which is not great for general compilation.

Signed-off-by: Pavel Emelyanov <ovzxemul@gmail.com>